+++ src/xrCore/LocatorAPI_auth.cpp
@@ -62,7 +62,7 @@ void CLocatorAPI::auth_runtime(void* params)
 
             // test for skip
             BOOL bSkip = FALSE;
-            for (u32 s = 0; s < _o->ignore.size(); s++)
+            for (size_t s = 0; s < _o->ignore.size(); s++)
             {
                 if (strstr(f.name, _o->ignore[s].c_str()))
                     bSkip = TRUE;
@@ -72,7 +72,7 @@ void CLocatorAPI::auth_runtime(void* params)
                 continue;
 
             // test for important
-            for (u32 s = 0; s < _o->important.size(); s++)
+            for (size_t s = 0; s < _o->important.size(); s++)
             {
                 if ((f.size_real != 0) && strstr(f.name, _o->important[s].c_str()))
                 {
